fix(build): resolve CI compilation errors across platforms
Fix several compilation errors that were blocking the full Descent3
link and all CI builds:

- LegacyCockpit.h: move tCockpitInfo/tCockpitCfgInfo struct definitions
  into the header so the value member m_info has a complete type
- config.cpp: cockpit_mode radio button pointer must be int* (not short*)
  to match AddFirstRadioButton's return type
- postrender.cpp: Postrender_sort_func comparator now takes const refs
  instead of pointers to work with std::sort
- damage.cpp: include ddio.h for timer_GetMSTime
- multi_server.cpp: include sounds.h for SOUND_NONE_INDEX
- WidescreenCockpit.cpp: include gauges.h for RenderGauges

Bug: https://github.com/DescentDevelopers/Descent3/issues/560

+++ Descent3/config.cpp
@@ -773,7 +773,7 @@ struct video_menu {
   bool *vsync = nullptr;
   char *resolution_string = nullptr;
   short *fov = nullptr;
-  short *cockpit_mode = nullptr;
+  int *cockpit_mode = nullptr;
   bool resolution_changed = false;
 
   int *bitdepth = nullptr; // bitdepths
